session/handler_item_stack_request.go: Handle regular crafting actions.

+++ server/session/session.go
@@ -304,7 +304,7 @@ func (s *Session) handlePackets() {
 }
 
 // craftingSize gets the crafting size based on the opened container ID.
-func (s *Session) craftingSize() byte {
+func (s *Session) craftingSize() uint32 {
 	if s.openedContainerID.Load() == 1 {
 		return craftingSizeLarge
 	}
@@ -312,7 +312,7 @@ func (s *Session) craftingSize() byte {
 }
 
 // craftingOffset gets the crafting offset based on the opened container ID.
-func (s *Session) craftingOffset() byte {
+func (s *Session) craftingOffset() uint32 {
 	if s.openedContainerID.Load() == 1 {
 		return craftingGridLargeOffset
 	}
